add category model
All of the work I completed is on one branch because I did not understand how to use GitKraken and didn't realize until it was too late that I should've made a branch for each of my Trello cards (assignments).  On this branch, I have set up the Categories and Projects models, created Categories and Projects controller and created routes for both Categories and Projects.